// src/app/Models/HomeDeliveryModel.ts
export interface HomeDeliveryModel {
  idHomeDelivery: number;
  idShop: number;
  logoType: string;
  shopName: string;
  idOffice: number;
  idSale: string;
  idStoreUser: any;
  document: string;
  idClient: string;
  status: string;
  address: string;
  pickupLocation: string;
  latitudeLocation: string;
  longitudeLocation: string;
  deliveryTimeCalculation: string;
  dateAssigment: string;
  datePartureDate: string;
  deliveryDate: string;
  dateCreated: string;
  dateUpdated: string;
}